Refactor into scratch based image
The previous setup wasn't compatible with non-AMD64 builder nodes,
as the published (multi-arch) image result always targeted the
platform defined during build time.

This commit refactors the Makefile and previous Dockerfile revision
into a Makefile recipe, which can be copied over as a prerequiste
step to compile and install `libgit2` with a predefined
configuration before the build of a Go application.

This makes the `Dockerfile` on the consumer side a bit fatter, but
allows to distribute the configuration while making cross
compiliation to and from multiple architectures still work.

Signed-off-by: Hidde Beydals <hello@hidde.co>